pyapi/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
reportlab
python-dotenv
